+++ src/Farmer/Builders/Builders.NetworkInterface.fs
@@ -39,6 +39,7 @@ type NetworkInterfaceConfig =
                         [
                             {
                                 SubnetName = subnet.Name
+                                ApplicationSecurityGroups = []
                                 LoadBalancerBackendAddressPools = []
                                 PublicIpAddress = None
                                 PrivateIpAllocation = Some(this.PrivateIpAddress)
@@ -83,6 +84,7 @@ type NetworkInterfaceConfig =
                             [
                                 {
                                     SubnetName = ResourceName subnetName
+                                    ApplicationSecurityGroups = []
                                     LoadBalancerBackendAddressPools = []
                                     PublicIpAddress = None
                                     PrivateIpAllocation = Some(this.PrivateIpAddress)