+++ src/Farmer/Builders/Builders.NetworkSecurityGroup.fs
@@ -7,6 +7,53 @@ open Farmer.Arm.NetworkSecurityGroup
 open Farmer.NetworkSecurity
 open System.Net
 
+/// Application security group configuration
+type ApplicationSecurityGroupConfig =
+    {
+        Name: ResourceName
+        Dependencies: ResourceId Set
+        Tags: Map<string, string>
+    }
+
+    interface IBuilder with
+        member this.BuildResources location =
+            [
+                {
+                    ApplicationSecurityGroup.Name = this.Name
+                    Location = location
+                    Dependencies = this.Dependencies
+                    Tags = this.Tags
+                }
+            ]
+
+        member this.ResourceId = applicationSecurityGroups.resourceId this.Name
+
+type ApplicationSecurityGroupBuilder() =
+    member _.Yield _ =
+        {
+            ApplicationSecurityGroupConfig.Name = ResourceName.Empty
+            Dependencies = Set.empty
+            Tags = Map.empty
+        }
+
+    [<CustomOperation "name">]
+    member _.Name(config: ApplicationSecurityGroupConfig, name) =
+        { config with Name = ResourceName name }
+
+    interface IDependable<ApplicationSecurityGroupConfig> with
+        member _.Add state newDeps =
+            { state with
+                Dependencies = state.Dependencies + newDeps
+            }
+
+    interface ITaggable<ApplicationSecurityGroupConfig> with
+        member _.Add state tags =
+            { state with
+                Tags = state.Tags |> Map.merge tags
+            }
+
+let applicationSecurityGroup = ApplicationSecurityGroupBuilder()
+
 /// Network access policy
 type SecurityRuleConfig =
     {
@@ -46,13 +93,34 @@ type SecurityRuleConfig =
             SourcePorts = this.Sources |> List.map (fun (_, _, sourcePort) -> sourcePort) |> Set
             SourceAddresses =
                 this.Sources
+                |> List.filter (fun (_, sourceEndpoint, _) ->
+                    match sourceEndpoint with
+                    | ApplicationSecurityGroup _ -> false
+                    | _ -> true)
                 |> List.map (fun (_, sourceAddress, _) -> sourceAddress)
                 |> List.distinct
+            SourceApplicationSecurityGroups =
+                this.Sources
+                |> List.choose (fun (_, sourceEndpoint, _) ->
+                    match sourceEndpoint with
+                    | ApplicationSecurityGroup asgId -> Some asgId
+                    | _ -> None)
             DestinationPorts =
                 match this.Services with
                 | [] -> Set [ AnyPort ]
                 | services -> services |> List.map (fun (NetworkService (_, port)) -> port) |> Set
-            DestinationAddresses = this.Destinations
+            DestinationAddresses =
+                this.Destinations
+                |> List.filter (fun endpoint ->
+                    match endpoint with
+                    | ApplicationSecurityGroup asgId -> false
+                    | _ -> true)
+            DestinationApplicationSecurityGroups =
+                this.Destinations
+                |> List.choose (fun endpoint ->
+                    match endpoint with
+                    | ApplicationSecurityGroup asgId -> Some asgId
+                    | _ -> None)
             Access = this.Operation
             Direction = this.Direction
             Priority = priority
@@ -151,6 +219,34 @@ type SecurityRuleBuilder() =
             Sources = (protocol, Host(IPAddress.Parse sourceAddress), AnyPort) :: state.Sources
         }
 
+    /// Sets the rule to a managed source application security group.
+    [<CustomOperation("add_source_application_security_group")>]
+    member _.AddSourceApplicationSecurityGroup
+        (
+            state: SecurityRuleConfig,
+            protocol,
+            asg: ApplicationSecurityGroupConfig
+        ) =
+        { state with
+            Sources =
+                (protocol, ApplicationSecurityGroup(Managed (asg :> IBuilder).ResourceId), AnyPort)
+                :: state.Sources
+        }
+
+    /// Sets the rule to an unmanaged source application security group.
+    [<CustomOperation("link_source_application_security_group")>]
+    member _.LinkSourceApplicationSecurityGroup
+        (
+            state: SecurityRuleConfig,
+            protocol,
+            asg: ApplicationSecurityGroupConfig
+        ) =
+        { state with
+            Sources =
+                (protocol, ApplicationSecurityGroup(Unmanaged (asg :> IBuilder).ResourceId), AnyPort)
+                :: state.Sources
+        }
+
     /// Sets the rule to match on a source network.
     [<CustomOperation("add_source_network")>]
     member _.AddSourceNetwork(state: SecurityRuleConfig, protocol, sourceNetwork) =
@@ -186,6 +282,29 @@ type SecurityRuleBuilder() =
             Destinations = Host(IPAddress.Parse destAddress) :: state.Destinations
         }
 
+    /// Sets the rule to a managed destination application security group.
+    [<CustomOperation("add_destination_application_security_group")>]
+    member _.AddDestinationApplicationSecurityGroup(state: SecurityRuleConfig, asg: ApplicationSecurityGroupConfig) =
+        { state with
+            Destinations =
+                ApplicationSecurityGroup(Managed (asg :> IBuilder).ResourceId)
+                :: state.Destinations
+        }
+
+    /// Sets the rule to an unmanaged destination application security group.
+    [<CustomOperation("link_destination_application_security_group")>]
+    member _.LinkDestinationApplicationSecurityGroup
+        (
+            state: SecurityRuleConfig,
+            protocol,
+            asg: ApplicationSecurityGroupConfig
+        ) =
+        { state with
+            Destinations =
+                ApplicationSecurityGroup(Unmanaged (asg :> IBuilder).ResourceId)
+                :: state.Destinations
+        }
+
     /// Sets the rule to match on a destination network.
     [<CustomOperation("add_destination_network")>]
     member _.AddDestinationNetwork(state: SecurityRuleConfig, destNetwork) =
@@ -214,6 +333,7 @@ let securityRule = SecurityRuleBuilder()
 type NsgConfig =
     {
         Name: ResourceName
+        Dependencies: ResourceId Set
         SecurityRules: SecurityRuleConfig list
         Tags: Map<string, string>
         InitialRulePriority: int
@@ -227,6 +347,7 @@ type NsgConfig =
             [
                 {
                     Name = this.Name
+                    Dependencies = this.Dependencies
                     Location = location
                     SecurityRules =
                         seq {
@@ -252,6 +373,7 @@ type NsgBuilder() =
     member _.Yield _ =
         {
             Name = ResourceName.Empty
+            Dependencies = Set.empty
             SecurityRules = []
             Tags = Map.empty
             InitialRulePriority = 100
@@ -282,6 +404,12 @@ type NsgBuilder() =
         { state with
             PriorityIncrementor = priority_incr
         }
+
+    interface IDependable<NsgConfig> with
+        member _.Add state newDeps =
+            { state with
+                Dependencies = state.Dependencies + newDeps
+            }
 
     interface ITaggable<NsgConfig> with
         member _.Add state tags =